gcs_to_bigquery/requirements.txt
fastapi
uvicorn
pandas
google-cloud-storage
google-cloud-bigquery
pyarrow
db-dtypes
